Added explicit casting.

+++ src/Drawer/ProgressScaler.h
@@ -17,22 +17,22 @@ namespace Egfx
 
 		static constexpr const uint16_t GetProgress(const uint32_t periodMicros, const uint32_t frameTime)
 		{
-			return ((uint64_t)(frameTime % periodMicros) * UINT16_MAX) / periodMicros;
+			return uint16_t(((uint64_t)(frameTime % periodMicros) * UINT16_MAX) / periodMicros);
 		}
 
 		static constexpr uint8_t ScaleProgress(const uint16_t progress, const uint8_t value)
 		{
-			return ScaleProgress(progress, (uint16_t)value);
+			return uint8_t(ScaleProgress(progress, (uint16_t)value));
 		}
 
 		static constexpr uint16_t ScaleProgress(const uint16_t progress, const uint16_t value)
 		{
-			return (((uint32_t)progress) * value) / UINT16_MAX;
+			return uint16_t((((uint32_t)progress) * value) / UINT16_MAX);
 		}
 
 		static constexpr uint16_t ScaleProgress(const uint16_t progress, const uint32_t value)
 		{
-			return (((uint64_t)progress) * value) / UINT16_MAX;
+			return uint16_t((((uint64_t)progress) * value) / UINT16_MAX);
 		}
 
 		static const uint16_t TriangleResponse(const uint16_t progress)
